feat: 메인 브금 사운드 추가 및 타이틀 씬 UI 구현

+++ Assets/Scripts/UI/TitleSceneManager.cs
@@ -1,3 +1,4 @@
+using System.Collections;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
@@ -7,25 +8,45 @@ public class TitleSceneManager : MonoBehaviour
     [SerializeField]
     private Button continueButton;
 
+    [SerializeField]
+    private AudioSource bgmSource;
+
+    [SerializeField]
+    private AudioSource sfxSource;
+
+    [SerializeField]
+    private AudioClip buttonClick;
+
     private void Start()
     {
         SaveLoadManager.Init();
         continueButton.interactable = SaveLoadManager.HasSaveData();
+        bgmSource.Play();
     }
 
     public void OnContinueButton()
     {
-        SceneManager.LoadScene("MainGameScene");
+        StartCoroutine(LoadScene("MainGameScene"));
     }
 
     public void OnNewGameButton()
     {
-        SaveLoadManager.DeleteSave();
-        SceneManager.LoadScene("MainGameScene");
+        StartCoroutine(LoadScene("MainGameScene", deleteSave: true));
     }
 
     public void OnQuitButton()
     {
+        sfxSource.PlayOneShot(buttonClick);
         Application.Quit();
+    }
+
+    private IEnumerator LoadScene(string sceneName, bool deleteSave = false)
+    {
+        sfxSource.PlayOneShot(buttonClick);
+        bgmSource.Stop();
+        yield return new WaitForSeconds(buttonClick.length);
+        if (deleteSave)
+            SaveLoadManager.DeleteSave();
+        SceneManager.LoadScene(sceneName);
     }
 }
